fix utils
