+++ backend/src/api/v1/beats.py
@@ -11,8 +11,9 @@ from src.schemas.beats import (
     SBeatDeleteResponse,
 )
 from src.services.beats import BeatsRepository
-from src.utils.beats import unique_filename
 from src.utils.auth import get_current_user
+from src.utils.files import unique_filename
+
 
 beats = APIRouter(prefix="/beats", tags=["Beats"])
 
@@ -24,7 +25,7 @@ beats = APIRouter(prefix="/beats", tags=["Beats"])
     responses={status.HTTP_200_OK: {"model": List[SBeatResponse]}},
 )
 async def get_user_beats(
-    user: SUser = Depends(get_current_user),
+        user: SUser = Depends(get_current_user),
 ) -> List[SBeatResponse]:
     response = await BeatsRepository.find_all(user=user)
 
@@ -60,7 +61,7 @@ async def get_one_beat(beat_id: int) -> SBeatResponse:
     responses={status.HTTP_200_OK: {"model": SBeatResponse}},
 )
 async def add_beats(
-    file: UploadFile = File(...), user: SUser = Depends(get_current_user)
+        file: UploadFile = File(...), user: SUser = Depends(get_current_user)
 ) -> SBeatResponse:
     file_info = await unique_filename(file) if file else None
     file_url = await MediaRepository.upload_file("AUDIOFILES", file_info, file)
@@ -85,7 +86,7 @@ async def add_beats(
     responses={status.HTTP_200_OK: {"model": SBeatResponse}},
 )
 async def update_pic_beats(
-    beats_id: int, file: UploadFile = File(...), user: SUser = Depends(get_current_user)
+        beats_id: int, file: UploadFile = File(...), user: SUser = Depends(get_current_user)
 ) -> SBeatResponse:
     file_info = await unique_filename(file) if file else None
     file_url = await MediaRepository.upload_file("PICTURES", file_info, file)
@@ -104,7 +105,7 @@ async def update_pic_beats(
     responses={status.HTTP_200_OK: {"model": SBeatResponse}},
 )
 async def release_beats(
-    beat_id: int, data: SBeatRelease, user: SUser = Depends(get_current_user)
+        beat_id: int, data: SBeatRelease, user: SUser = Depends(get_current_user)
 ) -> SBeatResponse:
     update_data = {}
 
@@ -129,7 +130,7 @@ async def release_beats(
     responses={status.HTTP_200_OK: {"model": SBeatResponse}},
 )
 async def update_beats(
-    beat_id: int, data: SBeatUpdate, user: SUser = Depends(get_current_user)
+        beat_id: int, data: SBeatUpdate, user: SUser = Depends(get_current_user)
 ) -> SBeatResponse:
     update_data = {}
 
@@ -156,7 +157,7 @@ async def update_beats(
     responses={status.HTTP_200_OK: {"model": SBeatDeleteResponse}},
 )
 async def delete_beats(
-    beat_id: int, user: SUser = Depends(get_current_user)
+        beat_id: int, user: SUser = Depends(get_current_user)
 ) -> SBeatDeleteResponse:
     await BeatsRepository.delete(id_=beat_id)
     return SBeatDeleteResponse
