+++ backend/src/api/v1/comments.py
@@ -18,6 +18,7 @@ from src.schemas.comments import (
 )
 from src.utils.auth import get_current_user
 
+
 comments = APIRouter(prefix="/comments", tags=["Comments"])
